Move NotAvailable to voir.instruments.gpu.common
